Cleanup: laziness, single lookup, file split for the presentation assembler
Follow-up polish on the #1907 extraction (no behaviour change):

- Restore the laziness the extraction lost. SelectedTripRenderInput built its
  directionUnderlay (a per-segment RoutePolyline copy) and stopPresentation
  (a Location stop projection) eagerly on every publish with a selected
  vehicle, but the assembler discards the underlay whenever the route has an
  adjacency colour and the stop projection whenever the trip is undrawable.
  The pre-extraction code computed both only inside the branch that used them.
  Defer both as thunks (mirroring the existing projectedFocusStops thunk) so
  each is computed only when its branch is taken.
- Collapse the selected-vehicle branch to one adjacency-colour lookup and drop
  the redundant `selected != null` guard (selectedTrip is non-null only when
  selected is), restoring the original single-nullable shape.
- Drop the now-dead `= { emptyMap() }` default on projectedFocusStops — the
  sole production caller always supplies it; make the contract explicit.
- Move the Location-dependent projection kernels (projectFocusedStops,
  projectStopsOntoPolylines, the shared nearest-point kernel) out of
  RouteMapPresentationPlan.kt into a new RouteStopProjection.kt, matching the
  instrumented RouteStopProjectionTest's name and leaving the assembler file
  cleanly JVM-pure.
